implement reset password logic

+++ routes/api/appartment.js
@@ -2,21 +2,21 @@ const express = require('express');
 const router = express.Router();
 const { check, validationResult } = require('express-validator');
 const Appartment = require('../../models/Appartment');
-const authorize = require('../../middlewares/auth');
+const { authanticate, authorize } = require('../../middlewares/auth');
 
 // get all appartments
 router.get('/', async (req, res) => {
 	try {
 		let appartments = await Appartment.find();
 
-		res.status(200).json({
+		return res.status(200).json({
 			status: 'success',
 			result: appartments.length,
 			data: appartments
 		});
 	} catch (error) {
 		console.log(error);
-		res.status(500).json({
+		return res.status(500).json({
 			status: 'Failed',
 			error
 		});
@@ -36,14 +36,14 @@ router.get('/:id', async (req, res) => {
 			});
 		}
 
-		res.status(200).json({
+		return res.status(200).json({
 			status: 'success',
 			result: appartment.length,
 			data: appartment
 		});
 	} catch (error) {
 		console.log(error);
-		res.status(400).json({
+		return res.status(400).json({
 			status: 'Failed',
 			error
 		});
@@ -51,20 +51,20 @@ router.get('/:id', async (req, res) => {
 });
 
 //get all appartment belonging to one user
-router.get('/user-appartments/:userID', authorize, async (req, res) => {
+router.get('/user-appartments/:userID', authanticate, async (req, res) => {
 	try {
 		let userAppartments = await Appartment.find({
 			'user.id': req.params.userID
 		});
 
-		res.status(200).json({
+		return res.status(200).json({
 			status: 'success',
 			result: userAppartments.length,
 			data: userAppartments
 		});
 	} catch (error) {
 		console.log(error);
-		res.status(400).json({
+		return res.status(400).json({
 			status: 'Failed',
 			error
 		});
@@ -72,9 +72,9 @@ router.get('/user-appartments/:userID', authorize, async (req, res) => {
 });
 
 //add an appartment
-router.post('/', [], authorize, async (req, res) => {
+router.post('/', [], authanticate, async (req, res) => {
 	console.log('add appartments route');
-	res.status(200).json({
+	return res.status(200).json({
 		status: 'success',
 		result: 1,
 		data: 'add appartment route'
@@ -82,7 +82,7 @@ router.post('/', [], authorize, async (req, res) => {
 });
 
 //modify an appartment
-router.patch('/:id', authorize, async (req, res) => {
+router.patch('/:id', authanticate, async (req, res) => {
 	try {
 		let updatedAppartment = await Appartment.findByIdAndUpdate(
 			req.params.id,
@@ -90,36 +90,41 @@ router.patch('/:id', authorize, async (req, res) => {
 			{ new: true, runValidators: true }
 		);
 
-		res.status(200).json({
+		return res.status(200).json({
 			status: 'success',
 			result: updatedAppartment.length,
 			data: updatedAppartment
 		});
 	} catch (error) {
 		console.log(error);
-		res.status(400).json({
+		return res.status(400).json({
 			status: 'Failed',
 			error
 		});
 	}
 });
 
-//delete an appartment
-router.delete('/:id', authorize, async (req, res) => {
-	try {
-		await Appartment.findByIdAndDelete(req.params.id);
+//delete an appartment, , restricted to admins developers and  users
+router.delete(
+	'/:id',
+	authanticate,
+	authorize('admin', 'developer', 'user'),
+	async (req, res) => {
+		try {
+			await Appartment.findByIdAndDelete(req.params.id);
 
-		res.status(204).json({
-			status: 'success',
-			message: 'Appartment deleted !'
-		});
-	} catch (error) {
-		console.log(error);
-		res.status(400).json({
-			status: 'Failed',
-			error
-		});
+			return res.status(204).json({
+				status: 'success',
+				message: 'Appartment deleted !'
+			});
+		} catch (error) {
+			console.log(error);
+			return res.status(400).json({
+				status: 'Failed',
+				error
+			});
+		}
 	}
-});
+);
 
 module.exports = router;
